debug output

+++ pyobs_ascom/ascomtelescope.py
@@ -107,7 +107,7 @@ class AscomTelescope(BaseTelescope, IFitsHeaderProvider, IEquatorialMount):
         log.info('cur_dec: %.5f', cur_dec)
         log.info('radians cur_dec: %.5f', np.radians(cur_dec))
         log.info('cos radians cur_dec: %.5f', np.cos(np.radians(cur_dec)))
-        log.info('ra offset: %.5f" -> %.5f"', self._offset_ra * 3600, self._offset_ra * np.cos(np.radians(cur_dec)) * 3600)
+        log.info('ra offset: %.5f" -> %.5f"', float(self._offset_ra * 3600), float(self._offset_ra * np.cos(np.radians(cur_dec)) * 3600))
         ra += float(self._offset_ra * np.cos(np.radians(cur_dec)))
         dec += float(self._offset_dec)
